desde 42

+++ libft/ft_strnstr.c
@@ -13,9 +13,9 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int	i;
-	int	j;
-	int k;
+	size_t	i;
+	size_t	j;
+	size_t k;
 
 	i = 0;
 	k = 0;
